fix: Re-raise SIGINT and SIGTERM (#562)
The signal handlers registered in the init function prevent Go's default
behavior (exiting with an exit code which depends on the OS and/or signal)
when receiving such a signal.

This commit re-raises the signal after handling it, triggering the
default behavior unless the consumer registered it's own handlers.

Note that this could lead to other unexpected behavior: Any other handlers
registered by the consumer via signal.Notify receive the same signal
twice, because the Go runtime sends each signal to all handlers registered
for that signal.